moved gRPC chat service to separate file
